+++ packages/jm-ms-core/bench/router.js
@@ -1,21 +1,21 @@
 const benchmark = require('benchmark')
 const suite = new benchmark.Suite()
 
-let Router = require('../lib/router')
+const Router = require('../lib/router')
 
-const handle1 = opts => {}
+const handle1 = () => {}
 
-const handle2 = opts => {
+const handle2 = () => {
   throw new Error('err')
 }
 
-const handle = opts => {
+const handle = () => {
   return { ret: 1 }
 }
 
-let r1 = new Router()
+const r1 = new Router()
 r1.use(handle1, handle1, handle)
-let r2 = new Router()
+const r2 = new Router()
 r2.use(handle1, handle2, handle)
 
 suite
@@ -23,13 +23,13 @@ suite
     r1
       .request('/')
       .then(() => {})
-      .catch(e => {})
+      .catch(() => {})
   })
   .add('request, with error', () => {
     r2
       .request('/')
       .then(() => {})
-      .catch(e => {})
+      .catch(() => {})
   })
   .on('cycle', function (event) {
     console.log(String(event.target))