+++ packages/jm-ms-core/lib/consts.js
@@ -1,0 +1,3 @@
+module.exports = {
+  Types: ['get', 'post', 'put', 'delete']
+}